+++ resources/js/Pages/Media.jsx
@@ -1,0 +1,10 @@
+import AuthenticatedLayout from "@/Layouts/AuthenticatedLayout";
+
+export default function Media() {
+  return (
+    <AuthenticatedLayout>
+      <h1>Media </h1>
+    </AuthenticatedLayout>
+
+  )
+}